chore: Move configuration to own package
* Card ID: CCT-1583

To improve code separation, move the configuration object to its own
package.

Generated-By: Claude (claude-sonnet-4)
